Factorized Embeddings method

Description

Factorized Embeddings is described in this article by Trofimov et al. 2020. This model is a self-supervised deep neural network that uses tensor factorization to simultaneously learn gene and sample representation spaces. This type of network is tailored to work with large RNA sequencing data, but can be applied to any type of large multivariate data. The FE model treats both factors (typically samples and gene) as factors contributing to characterizing the values (ie. gene expression data). This architecture generates sample representations that can be used for auxiliary tasks such as visualization or classification.

The code in this package is based on the first Factorized Embeddings analysis scripts according to the code provided by Trofimov et al. (2020) GitHub link. The demonstrated functionalities of Factorized Embeddings are implemented in the Julia programming language julialang.org using the Flux library Innes 2018. The main features of the code include: • Generation of low-dimensional sample embeddings that can be used for auxiliary tasks like 2D visualization or classification. • Highly configurable models via a dictionary of hyperparameters provided by the user. • During network optimization, models, learning curves, reconstruction performance, and, if possible, 2D sample embedding visualizations are recorded. • Once trained, the model can infer new points. • A functionality is also available for imputing the sample embedding space in 2D.

Model optimizations were tested using SGD with the ADAM optimizer (Kingma and Ba 2017) on GPUs (V100-SMX2, 32 GB) on servers with 64 GB RAM. It is important to note that this implementation requires access to a GPU for optimal performance. The code can be adapted for specific server constraints, such as reducing mini-batch sizes and limiting the number of network parameters. Details about the required package installations for GPU usage are also provided in this guide.

Schematic of Factorized Embeddings

Public Interface

Main methods

FactorizedEmbeddings.generate_params — Function

generate_params(X_data::AbstractArray; 
                emb_size::Int=2, emb_size_2::Int=100, 
                nsteps_dim_redux::Int=1000, l2_val::Float64=1e-7, 
                fe_layers_size = [100, 50, 50]
           )

Function that takes input hyper-parameters and outputs a dictonary.

FactorizedEmbeddings.fit — Function

fit(X_data; dim_redux_size::Int=2, nsteps::Int=1000, l2::Float64=1e-7)

This function instanciates a Factorized Embeddings model with default or imputed parameters. Then trains the model on the input data and returns the trained model.

fit(X_data, FE_params::Dict)

This function instanciates a Factorized Embeddings model imputed hyper-parameter dictionary. Then trains the model on the input data and returns the trained model.

FactorizedEmbeddings.fit_transform — Function

fit_transform(X_data; dim_redux_size::Int=2, nsteps::Int=1000, l2::Float64=1e-7, verbose::Int = 0)

This function instanciates a Factorized Embeddings model with default or imputed parameters. Then trains the model on the input data and returns the dimensionality-reduced sample embedding.

fit_transform(X_data, FE_params::Dict;verbose::Int=0)

This function instanciates a Factorized Embeddings model imputed hyper-parameter dictionary. Then trains the model on the input data and returns the dimensionality-reduced sample embedding.

FactorizedEmbeddings.infer — Function

infer(trainedFE, traindata, testdata, paramsdict;verbose=0)

Infers new data with the pre-trained model. Input parameters. trainedFE: the pre-trained Flux DNN model. traindata: the training dataset. testdata: the test dataset. paramsdict: Dictionary of hyper-parameters that was set during the training phase.

Other utility methods

FactorizedEmbeddings.fit! — Function

fit!(X_data, model, FE_params::Dict)

This function uses the inputed Factorized Embeddings model with pre-defined and imputed hyper-parameter dictionary. Then trains the model on the input data and returns the trained model.

FactorizedEmbeddings.reset_embedding_layer — Function

function reset_embedding_layer!(FE_net, new_embed; cp_dev=gpu)

This methods copies the imputed model and returns a model with inputed pre-defined embedding layer.

This function allows to initialise the sample embedding with a certain pre-defined embedding.

FactorizedEmbeddings.train! — Function

function train!(params, X, Y, model;verbose = 0)

This methods trains a model with the X and Y training data and returns the trained model.

FactorizedEmbeddings.FE_model — Function

FE_model(params::Dict)

This method takes as input a dictionary of hyperparameters instantiates and returns a Flux.Chain type DNN model. This model can then be fed to the train! method to fit the model's parameters to the data.

FactorizedEmbeddings.prep_FE — Function

prep_FE(data::Matrix, device=gpu; order = "shuffled", verbose = 0)

This method takes the data matrix as input and outputs the vectors X containing two vectors of samples indices and gene indices and Y containing the expression values of the sample and the gene.