Factorized Embeddings method

Main methods

FactorizedEmbeddings.generate_params — Function

generate_params(X_data::AbstractArray; 
                emb_size::Int=2, emb_size_2::Int=100, 
                nsteps_dim_redux::Int=1000, l2_val::Float64=1e-7, 
                fe_layers_size = [100, 50, 50]
           )

Function that takes input hyper-parameters and outputs a dictonary.

FactorizedEmbeddings.fit — Function

fit(X_data; dim_redux_size::Int=2, nsteps::Int=1000, l2::Float64=1e-7)

This function instanciates a Factorized Embeddings model with default or imputed parameters. Then trains the model on the input data and returns the trained model.

fit(X_data, FE_params::Dict)

This function instanciates a Factorized Embeddings model imputed hyper-parameter dictionary. Then trains the model on the input data and returns the trained model.

FactorizedEmbeddings.fit_transform — Function

fit_transform(X_data; FE_params::Dict)

This function instanciates a Factorized Embeddings model imputed hyper-parameter dictionary. Then trains the model on the input data and returns the dimensionality-reduced sample embedding.

FactorizedEmbeddings.infer — Function

infer(trainedFE, traindata, trainids, testdata, testids, samples, genes, paramsdict)

Infers new data with the pre-trained model.